shape_osm/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
GeoAlchemy2==0.4.0
idna==2.6
numpy==1.13.1
pandas==0.20.3
psycopg2==2.7.3
pyaml==17.8.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
Shapely==1.6.0
six==1.10.0
SQLAlchemy==1.1.13
SQLAlchemy-Utils==0.32.14
